[refactor] `sam-highlighter` -> `sam-react-common`
The goal is to have a centralized place to store common libraries
