+++ .github/workflows/generated-cd-blog.yml
@@ -9,7 +9,7 @@ on:
       - .github/workflows/generated-cd-blog.yml
       - package.json
       - 'configuration/**'
-      - 'sam-highlighter/**'
+      - 'sam-react-common/**'
       - 'blog/**'
 
 jobs:
