+++ .github/workflows/generated-cd-main-site-frontend.yml
@@ -9,7 +9,7 @@ on:
       - .github/workflows/generated-cd-main-site-frontend.yml
       - package.json
       - 'configuration/**'
-      - 'sam-highlighter/**'
+      - 'sam-react-common/**'
       - 'main-site-frontend/**'
 
 jobs:
