+++ .github/workflows/generated-cd-samlang-demo-frontend.yml
@@ -9,7 +9,7 @@ on:
       - .github/workflows/generated-cd-samlang-demo-frontend.yml
       - package.json
       - 'configuration/**'
-      - 'sam-highlighter/**'
+      - 'sam-react-common/**'
       - 'samlang-demo-frontend/**'
 
 jobs:
