+++ .github/workflows/generated-ci-main-site-frontend.yml
@@ -7,7 +7,7 @@ on:
       - .github/workflows/generated-ci-main-site-frontend.yml
       - package.json
       - 'configuration/**'
-      - 'sam-highlighter/**'
+      - 'sam-react-common/**'
       - 'main-site-frontend/**'
 
 jobs:
