+++ .github/workflows/generated-ci-samlang-demo-frontend.yml
@@ -7,7 +7,7 @@ on:
       - .github/workflows/generated-ci-samlang-demo-frontend.yml
       - package.json
       - 'configuration/**'
-      - 'sam-highlighter/**'
+      - 'sam-react-common/**'
       - 'samlang-demo-frontend/**'
 
 jobs:
