+++ main-site-frontend/package.json
@@ -6,7 +6,7 @@
     "@material-ui/core": "^4.3.3",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
-    "sam-highlighter": "0.1.0"
+    "sam-react-common": "0.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
