+++ sam-react-common/tsconfig.json
