+++ samlang-demo-frontend/src/LanguageDemo/ResultCard.tsx
@@ -2,8 +2,8 @@ import React, { ReactNode, ReactElement } from 'react';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import CardHeader from '@material-ui/core/CardHeader';
-import CodeBlock from 'sam-highlighter/CodeBlock';
-import initialize from 'sam-highlighter/language';
+import CodeBlock from 'sam-react-common/CodeBlock';
+import initialize from 'sam-react-common/language';
 import { Response } from './interpret';
 import DemoStyles from './LanguageDemo.module.css';
 import ResultStyles from './ResultCard.module.css';
